+++ examples/react-native/expo-app/App.tsx
@@ -1,14 +1,13 @@
 import {StatusBar} from 'expo-status-bar';
 import {StyleSheet, Text, View} from 'react-native';
 import {useEffect} from 'react';
-import {identify, Identify, init, track} from '@amplitude/analytics-react-native';
-import {LogLevel} from '@amplitude/analytics-types';
+import {identify, Identify, init, track, Types} from 'amplitude-rn-analytics';
 
 export default function App() {
   useEffect(() => {
     (async () => {
         await init('API_KEY', 'example_user_id', {
-            logLevel: LogLevel.Verbose,
+            logLevel: Types.LogLevel.Verbose,
         }).promise;
         track('test');
         await identify(new Identify().set('react-native-test', 'yes')).promise;